Add hooks for before and after the container is created.
The hooks are:

/etc/lxc-debian/host-pre-create
 - run on the host before the container is created

/etc/lxc-debian/host-post-create
 - run on the host after the container is created

/etc/lxc-debian/guest-post-create
 - run inside the guest (by chroot) after the the container is created

They must be executable to be run.

The host hooks are sourced, which means they have access to the variables in
the lxc-debian script, e.g. $NAME for the container name.

The guest hook is run by chrooting to the rootfs and calling /bin/bash on it.