Halts a linux container.
One downside: you have to specify the command to run to halt the
container. This is because there's no nice way to connect to a
container just to run one command yet.

Once you do specify the command, the script will take care of issuing
the halt command, and lxc-stopping the container once the machine is
fully halted.